% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/research_divisions.R
\name{addDivisions}
\alias{addDivisions}
\title{Append columns of research division dummy variables to the citations
 data frame.}
\usage{
addDivisions(cites_df)
}
\arguments{
\item{cites_df}{The citations data frame.}
}
\value{
The original citations dataframe with the division columns added.
}
\description{
One column is added for each division represented by any article in
the citations data frame. The columns have 0/1 values, indicating which
articles are in that research division.
}

