% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bucket_analysis.R
\name{compareByImpactFactorBuckets}
\alias{compareByImpactFactorBuckets}
\title{Compare on- and off-Academia citations within years and quantile groups
of journal impact factors.}
\usage{
compareByImpactFactorBuckets(cites_df, summarizer = mean, comparator = `/`,
  ...)
}
\arguments{
\item{cites_df}{A dataframe with citations and impact factors.}

\item{summarizer}{(default mean) A function to summarize citations within groups.}

\item{comparator}{(default `/`) A function with arguments (on, off), that compares
on and off-Academia citation summaries. The default computes the on/off ratio.}

\item{...}{Extra parameters to `quantileBuckets`}
}
\value{
A dataframe with statistic by year, impact factor group, and on/off-source.
}
\description{
Compare on- and off-Academia citations within years and quantile groups
of journal impact factors.
}

