% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_data.R
\name{importData}
\alias{importData}
\title{Import and combine paper citations, paper downloads, and journal information.}
\usage{
importData(cites_path = NULL, journals_path = NULL, doc_types_path = NULL,
  ...)
}
\arguments{
\item{cites_path}{A path to a file with paper citation data for papers both
on and off Academia,}

\item{journals_path}{A path to a file with data on academic journals.}

\item{...}{Other arguments to import and join functions.}

\item{doc_types_dirpath}{A path to a *directory* containing one or more files with
document categorization results (Original Research, Commentary/Editorial, etc.)}
}
\value{
A dataframe of citation, download, and journal data merged together.
}
\description{
Import and combine paper citations, paper downloads, and journal information.
}

