% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/figures.R
\name{makeFigure}
\alias{makeFigure}
\title{Function to generate figures from the paper.}
\usage{
makeFigure(n, cites_df, ...)
}
\arguments{
\item{n}{Figure caption number.}

\item{cites_df}{A data frame with article citations and journal data, as produced by `importData`.}

\item{...}{Optional arguments passed to figure functions.}
}
\value{
Nothing. Renders a plot.
}
\description{
Recreate a figure with a given citations dataset by specifying the table's
caption number in the paper.
}

