% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_r_models.R
\name{makeFormula}
\alias{makeFormula}
\title{Create a formula describing a regression of citations on covariates.}
\usage{
makeFormula(df, model = c("linear", "logistic", "negbin", "zinb"),
  division_interactions = TRUE)
}
\arguments{
\item{df}{A dataframe with citations data, and divisions as created by
`importData` followed by `addDivisions`}

\item{model}{One of 'linear', 'logistic', 'negbin', or 'zinb', the last for
zero-inflated negative binomial model.}

\item{division_interactions}{If TRUE (default), interact each division dummy
with the on-Academia dummy.}
}
\value{
An R formula.
}
\description{
Creates a formula of the form `y ~ ...` where `y` is the citation
count variable, possibly transformed, and `...` are covariates and their
interactions.
}
\details{
The linear model log-transforms the citations variable, while the negative
binomial and ZINB models do not. The logistic model transforms citations
into a 0/1 variable where 1 indicates the article received at least one
citation.
}

