% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables.R
\name{makeTable}
\alias{makeTable}
\title{Function to generate tables from the paper.}
\usage{
makeTable(n, cites_df, ...)
}
\arguments{
\item{n}{Table caption number.}

\item{cites_df}{A data frame with article citations and journal data, as produced by `importData`.}

\item{...}{Optional arguments passed to table functions.}
}
\value{
Nothing. Prints a table, usually in markdown, but possibly plain text or LaTeX.
}
\description{
Recreate a table, in markdown format with a given citations
dataset by specifying the table's caption number in the paper.
}

