% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_melting.R
\name{meltOnSets}
\alias{meltOnSets}
\title{Repeat rows of a dataframe for each element in a set.}
\usage{
meltOnSets(df, col, prefix, ...)
}
\arguments{
\item{df}{A dataframe}

\item{col}{A column in dataframe that is a character vector containing string
representations of sets.}

\item{prefix}{A string to name the column of set elements}

\item{...}{Other arguments to pass to `setParse`.}
}
\value{
A 'molten' dataframe.
}
\description{
Repeat rows of a dataframe for each element in a set.
}
\examples{
mydf <- data.frame(x=c(1, 2), ids=c('{1,2,3}', '{5, 10}'))
meltOnSets(mydf, 'ids', prefix='id', cast_to='integer')
#    x     ids variable id
# 1  1 {1,2,3}     id_1  1
# 2  2 {5, 10}     id_1  5
# 3  1 {1,2,3}     id_2  2
# 4  2 {5, 10}     id_2 10
# 5  1 {1,2,3}     id_3  3
}

