% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/import_data.R
\name{paperAgeYears}
\alias{paperAgeYears}
\title{Estimate the age of a paper in years.}
\usage{
paperAgeYears(date_collected, published_year)
}
\arguments{
\item{date_collected}{A Date vector of dates when the articles citations
were recorded.}

\item{published_year}{An integer vector of articles' publication years.}
}
\value{
A numeric vector of years between publication and citation
collection for each article.
}
\description{
Since we only have data on publication year, we approximate an article's
publication date as June 30 of its publication year.
}
\examples{
colln_dates = as.Date(c('2014-06-19', '2014-05-18', '2014-07-03'))
pub_years = c(2010, 2011, 2012)
paperAgeYears(colln_dates, pub_years)
}

