% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bucket_analysis.R
\name{plotByImpactFactorBuckets}
\alias{plotByImpactFactorBuckets}
\title{Boxplots on- and off-Academia citations within years and quantile groups
of journal impact factors.}
\usage{
plotByImpactFactorBuckets(cites_df, ...)
}
\arguments{
\item{cites_df}{A dataframe with citations and impact factors.}

\item{...}{Extra parameters to `quantileBuckets`}
}
\value{
A ggplot2 plot.
}
\description{
Boxplots on- and off-Academia citations within years and quantile groups
of journal impact factors.
}

