% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plotLogScale}
\alias{plotLogScale}
\title{Scale a plot's axis or axes using a x->log(1+x) transformation.}
\usage{
plotLogScale(plot, xy = c("x", "y", c("x", "y")))
}
\arguments{
\item{plot}{A ggplot2 plot with a defined aesthetic mapping}

\item{xy}{Axes to scale: 'x', 'y' or c('x', 'y')}
}
\value{
A new ggplot2 plot with scaled axes. The axis labels
will still reflect the raw data.
}
\description{
Scale a plot's axis or axes using a x->log(1+x) transformation.
}

