% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bucket_analysis.R
\name{quantileBuckets}
\alias{quantileBuckets}
\title{Group a variable into buckets based on its quantiles or those of another
variable.}
\usage{
quantileBuckets(x_quantile, x_bucket = x_quantile, nbuckets = 10,
  probs = NULL)
}
\arguments{
\item{x_quantile}{The variable to calculate quantile buckets from.}

\item{x_bucket}{The variable to collect into the quantile buckets.}

\item{nbuckets}{The number of quantile buckets to use. Specify this *or*
`probs`, but not both.}

\item{probs}{The vector of probabilities for the quantile bucket cut points.
Specify this *or* `nbuckets`, but not both.}
}
\value{
A factor vector corresponding to `x_bucket` with bucket ranges.
If an element of `x_bucket` is outside of the range of `x_quantile`, its
bucket will be NA.
}
\description{
Group a variable into buckets based on its quantiles or those of another
variable.
}

