% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_r_models.R
\name{runLinearModel}
\alias{runLinearModel}
\title{Fit a linear model of citations using lm.}
\usage{
runLinearModel(df, ...)
}
\arguments{
\item{df}{A dataframe with citations data, and
divisions as created by `importData` followed by
`addDivisions`}

\item{...}{Extra arguments to `lm`}
}
\value{
An `lm` object.
}
\description{
Fit a linear model of citations using lm.
}

