% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_r_models.R
\name{runZeroInflNegBinGLM}
\alias{runZeroInflNegBinGLM}
\title{Fit a Zero Inflated Negative Binomial model of
citations.}
\usage{
runZeroInflNegBinGLM(df, ...)
}
\arguments{
\item{df}{A dataframe with citations data, and
divisions as created by `importData` followed by
`addDivisions`}

\item{...}{Extra arguments to `pscl::zeroinfl`}
}
\value{
A fitted `zeroinfl` object.
}
\description{
Uses `pscl::zeroinfl` to fit the model. For parameters initial guesses
we use the coefficients from a logistic regression
(for the 'zero' parameters) and from a negative binomial regression
(for the 'count' parameters).
}

