% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/research_divisions.R
\name{selectDivisionCols}
\alias{selectDivisionCols}
\title{Select the FoR division data columns from a data frame}
\usage{
selectDivisionCols(df, pattern = "Div\\\\.")
}
\arguments{
\item{df}{A dataframe with division columns.}

\item{pattern}{(default "^div\\.") A regex pattern
for detecting division column names.}
}
\value{
A dataframe of only the division columns.
}
\description{
Select the FoR division data columns from a data frame
}

