% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_melting.R
\name{setParse}
\alias{setParse}
\title{Parse a vectors of string representations of sets to a list of vectors}
\usage{
setParse(set_str, cast_to = NULL, prefix = "v")
}
\arguments{
\item{set_str}{A character vector of string representations of sets.}

\item{cast_to}{The name of a data type to cast the elements of the set to.}

\item{prefix}{A prefix to name columns in the data frame (default 'v').}
}
\value{
A data frame with columns equal to the largest set's cardinality.
}
\description{
Parse a vectors of string representations of sets to a list of vectors
}
\examples{
setParse(c('{1, 2, 3}', '{2, 3}'), 'integer')
# v1 v2 v3
# 1  1  2  3
# 2  2  3 NA
setParse(c('{a, b}', '{}'))
#   v1 v2
# 1  a  b
}

