% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/bucket_analysis.R
\name{summarizeOverBuckets}
\alias{summarizeOverBuckets}
\title{Average results over buckets, weighting by the number of on-Academia
articles in the bucket.}
\usage{
summarizeOverBuckets(cites_df)
}
\arguments{
\item{cites_df}{A dataframe of citations with impact factors.}
}
\value{
A dataframe of weighted average results by year.
}
\description{
Average results over buckets, weighting by the number of on-Academia
articles in the bucket.
}

