% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/set_melting.R
\name{toSetStr}
\alias{toSetStr}
\title{Combine columns of a dataframe into a string set representation.}
\usage{
toSetStr(df)
}
\arguments{
\item{df}{A data frame whose columns will be combined into a set.}
}
\value{
A character vector of string representation forms.
}
\description{
Note that the uniqueness of the set is not ensured in this
function. If a row of the input data frame has the same data
in multiple columns, the resulting set representation will
have non-unique elements.
}
\examples{
myDF <- data.frame(x=c(1,2,3), y=c('a', 'b', 'c'))
toSetStr(myDF)  # {'1', 'a'} {'2', 'b'} {'3', 'c'}
}

