% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_r_models.R
\name{runLogisticModel}
\alias{runLogisticModel}
\title{Fit a Logistic model of citations using glm.}
\usage{
runLogisticModel(df, ...)
}
\arguments{
\item{df}{A dataframe with citations data, and
divisions as created by `importData` followed by
`addDivisions`}

\item{...}{Extra arguments to `glm`.}
}
\value{
A fitted GLM object.
}
\description{
Fit a Logistic model of citations using glm.
}

