% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/run_r_models.R
\name{runNegBinModel}
\alias{runNegBinModel}
\title{Fit a Negative Binomial model of citations using glm.}
\usage{
runNegBinModel(df, ...)
}
\arguments{
\item{df}{A dataframe with citations data, and
keyword factors as created by `importData` followed by
`addDivisions`}

\item{...}{Extra arguments to `MASS::glm.nb`}
}
\value{
A fitted GLM object.
}
\description{
Fit a Negative Binomial model of citations using glm.
}

